+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+* pkg: [nodejs](https://nodejs.org); `sync` and `update` install the latest version of Node.js on Linux and macOS
+
 ### Fixed
 
 * Windows: fix the build again :S